requirements.txt
nltk
pandas
numpy
python-Levenshtein